\name{twingp-package}
\alias{twingp-package}
\docType{package}
\title{A Global-Local Approximation Framework for Large-Scale Gaussian Process Modeling}
\description{
    For further details on the methodology, please refer to Vakayil and Joseph (2023). The package uses \code{nlopt} (Johnson, 2007) C++ library for hyperparameter optimization, \code{nanoflann} (Blanco and Rai, 2014) C++ library for nearest neighbor queries, and \code{Eigen} (Guennebaud and Jacob, 2010) C++ library for matrix operations.
}

\references{
	Vakayil, A., & Joseph, V. R. (2023). A Global-Local Approximation Framework for Large-Scale Gaussian Process Modeling. ArXiv [Stat.ML]. http://arxiv.org/abs/2305.10158

    Johnson, S. G. (2007), The NLopt nonlinear-optimization package. http://github.com/stevengj/nlopt

    Guennebaud, G., Jacob, B., & Others. (2010). Eigen v3. http://eigen.tuxfamily.org

    Blanco, J. L. & Rai, P. K. (2014). nanoflann: a C++ header-only fork of FLANN, a library for nearest neighbor (NN) with kd-trees. https://github.com/jlblancoc/nanoflann
}


